Router refactoring
In progress
